refactoring & ReSharper disabling
