+++ FileCommander/App.config
@@ -2,7 +2,7 @@
 <configuration>
     <appSettings>
         <add key="PromptDuplicitFileCopy" value="true" />
-        <add key="PromptDuplicitMoveCopy" value="true" />
+        <add key="PromptDuplicitFileMove" value="true" />
         <add key="PromptDeletion" value="true" />
     </appSettings>
 </configuration>
