+++ FileCommander/GUI/PromptUserDialogWindow.cs
@@ -1,9 +1,9 @@
 namespace FileCommander.GUI;
 using Gtk;
 
-public class UserPromptDialogWindow : Dialog
+public class PromptUserDialogWindow : Dialog
 {
-    public UserPromptDialogWindow(string prompt)
+    public PromptUserDialogWindow(string prompt)
     {
         MessageDialog md = new MessageDialog(this, 
             DialogFlags.DestroyWithParent, MessageType.Info,
